+++ include/PricerSolverSimpleDP.hpp
@@ -1,5 +1,6 @@
 #ifndef PRICER_SOLVER_SIMPLE_DP_HPP
 #define PRICER_SOLVER_SIMPLE_DP_HPP
+#include <cstddef>
 #include <memory>
 #include "Instance.h"
 #include "PricerSolverBase.hpp"
@@ -64,8 +65,8 @@ class PricerSolverSimpleDp : public PricerSolverBase {
     void   iterate_zdd() override;
     void   create_dot_zdd(const char* name) override;
     void   print_number_nodes_edges() override;
-    int    get_num_remove_nodes() override;
-    int    get_num_remove_edges() override;
+    size_t get_num_remove_nodes() override;
+    size_t get_num_remove_edges() override;
     size_t get_nb_edges() override;
     size_t get_nb_vertices() override;
     int    get_num_layers() override;
