+++ src/ModelInterface.cpp
@@ -13,6 +13,5 @@ ReformulationModel::ReformulationModel(size_t nb_assignments,
           }) |
           ranges::to<std::vector<std::shared_ptr<ConstraintBase>>>()) {
     auto m = static_cast<double>(nb_machines);
-    (*this).push_back(
-        std::make_shared<ConstraintConvex>(static_cast<double>(-m)));
+    (*this).push_back(std::make_shared<ConstraintConvex>(-m));
 }
